requirements.txt
gradio
phidata
google-generativeai
duckduckgo-search